--- crates/xwebtransport-core/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "xwebtransport-core"
-version = "0.2.0"
-edition = "2021"
-resolver = "2"
-license = "MIT"
-description = """
-A common WebTransport interface for browser and native.
-Write once, run anywhere. Originally intended for use with bevy.
-"""
-repository = "https://github.com/MOZGIII/xwebtransport"
-
-[dependencies]
-async-trait = "0.1"
-tokio = { version = "1", default-features = false, features = [] }
