--- crates/xwebtransport-error/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "xwebtransport-error"
-version = "0.2.0"
-edition = "2021"
-resolver = "2"
-license = "MIT"
-description = """
-A suite of reusable error types that naturally emerge from the xwebtransport
-API.
-Use when you don't want/need your own more precise types.
-"""
-repository = "https://github.com/MOZGIII/xwebtransport"
-
-[dependencies]
-xwebtransport-core = { version = "0.2", path = "../xwebtransport-core", default-features = false }
