--- crates/xwebtransport-tests/Cargo.toml
@@ -1,19 +1,0 @@
-[package]
-name = "xwebtransport-tests"
-version = "0.2.0"
-edition = "2021"
-resolver = "2"
-license = "MIT"
-description = """
-Common test suite for xwebtransport.
-Intended for reuse to provide the same coverage in all the execution
-environments.
-"""
-repository = "https://github.com/MOZGIII/xwebtransport"
-
-[dependencies]
-xwebtransport-core = { version = "0.2", path = "../xwebtransport-core", default-features = false }
-xwebtransport-error = { version = "0.2", path = "../xwebtransport-error", default-features = false }
-
-thiserror = "1"
-tokio = { version = "1", default-features = false, features = ["io-util"] }
