--- crates/xwebtransport-wtransport/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "xwebtransport-wtransport"
-version = "0.2.0"
-edition = "2021"
-resolver = "2"
-license = "MIT"
-description = """
-An implementation of the xwebtransport that runs natively via wtransport crate.
-"""
-repository = "https://github.com/MOZGIII/xwebtransport"
-
-[target.'cfg(not(target_family = "wasm"))'.dependencies]
-xwebtransport-core = { version = "0.2", path = "../xwebtransport-core", default-features = false }
-
-wtransport = { version = "0.1.8" }
-
-[target.'cfg(not(target_family = "wasm"))'.dev-dependencies]
-xwebtransport-tests = { version = "0.2", path = "../xwebtransport-tests" }
-
-anyhow = "1"
-tokio = { version = "1", features = ["macros"] }
