--- crates/xwebtransport/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "xwebtransport"
-version = "0.3.0"
-edition = "2021"
-resolver = "2"
-license = "MIT"
-description = """
-A common WebTransport interface for browser and native.
-Write once, run anywhere. Originally intended for use with bevy.
-"""
-repository = "https://github.com/MOZGIII/xwebtransport"
-
-[target.'cfg(target_family = "wasm")'.dependencies]
-xwebtransport-web-sys = { version = "0.3", path = "../xwebtransport-web-sys", default-features = false }
-
-[target.'cfg(not(target_family = "wasm"))'.dependencies]
-xwebtransport-wtransport = { version = "0.2", path = "../xwebtransport-wtransport", default-features = false }
