--- crates/xwebtransport/src/lib.rs
@@ -1,9 +1,0 @@
-#[cfg(target_family = "wasm")]
-pub use xwebtransport_web_sys as web_sys;
-#[cfg(not(target_family = "wasm"))]
-pub use xwebtransport_wtransport as wtransport;
-
-#[cfg(target_family = "wasm")]
-pub use self::web_sys as current;
-#[cfg(not(target_family = "wasm"))]
-pub use self::wtransport as current;
