+++ crates/xwt-core/src/datagram_utils.rs
