+++ crates/xwt-core/src/io_utils.rs
