+++ crates/xwt-core/src/trait_utils.rs
