+++ crates/xwt-core/src/traits.rs
