+++ crates/xwt-core/src/utils/dummy.rs
