+++ crates/xwt-error/src/impls.rs
