+++ crates/xwt-web-sys/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "xwt-web-sys"
+version = "0.3.1"
+edition = "2021"
+resolver = "2"
+license = "MIT"
+description = """
+An implementation of the xwt that runs in the browser.
+Powered by wasm-bindgen and web-sys.
+"""
+repository = "https://github.com/MOZGIII/xwt"
+
+[target.'cfg(target_family = "wasm")'.dependencies]
+xwt-core = { version = "0.2", path = "../xwt-core", default-features = false }
+web-sys-async-io = { version = "0.2", path = "../web-sys-async-io" }
+web-sys-stream-utils = { version = "0.2", path = "../web-sys-stream-utils" }
+
+tokio = { version = "1", default-features = false, features = [] }
+wasm-bindgen-futures = { version = "0.4" }
+wasm-bindgen = { version = "0.2" }
+js-sys = "0.3"
+web-sys = { version = "0.3", features = [
+  "WebTransport",
+  "WebTransportOptions",
+  "WebTransportBidirectionalStream",
+  "WebTransportSendStream",
+  "WebTransportReceiveStream",
+  "ReadableStreamDefaultReader",
+  "WritableStreamDefaultWriter",
+  "WebTransportDatagramDuplexStream",
+] }
+
+[target.'cfg(target_family = "wasm")'.dev-dependencies]
+xwt-tests = { version = "0.2", path = "../xwt-tests" }
+
+wasm-bindgen-test = "0.3"
