+++ crates/xwt-web-sys/src/error.rs
